@@ -13,48 +13,48 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@giscus/react": "^2.2.8",
-    "@tabler/icons-react": "^2.17.0",
-    "@vercel/analytics": "^1.0.0",
-    "class-variance-authority": "^0.5.3",
-    "date-fns": "^2.29.3",
-    "date-fns-tz": "^2.0.0",
+    "@giscus/react": "^3.0.0",
+    "@tabler/icons-react": "^3.5.0",
+    "@vercel/analytics": "^1.3.1",
+    "class-variance-authority": "^0.7.0",
+    "date-fns": "^3.6.0",
+    "date-fns-tz": "^3.1.3",
     "feed": "^4.2.2",
-    "katex": "^0.16.6",
-    "next": "13.3.1",
-    "next-mdx-remote": "^4.4.1",
-    "next-tweet": "^0.8.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-intersection-observer": "^9.4.3",
+    "katex": "^0.16.10",
+    "next": "14.2.3",
+    "next-mdx-remote": "^5.0.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-intersection-observer": "^9.10.2",
+    "react-tweet": "^3.2.1",
     "rough-notation": "^0.5.1"
   },
   "devDependencies": {
-    "@next/env": "13.3.1",
-    "@tailwindcss/typography": "^0.5.9",
-    "@types/hast": "^2.3.4",
-    "@types/node": "^18.16.1",
-    "@types/react": "18.2.0",
-    "@types/react-dom": "18.2.1",
-    "autoprefixer": "^10.4.14",
-    "eslint": "^8.39.0",
-    "eslint-config-next": "13.3.1",
-    "eslint-config-prettier": "^8.8.0",
+    "@next/env": "14.2.3",
+    "@tailwindcss/typography": "^0.5.13",
+    "@types/hast": "^3.0.4",
+    "@types/node": "^20.12.12",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
+    "autoprefixer": "^10.4.19",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "14.2.3",
+    "eslint-config-prettier": "^9.1.0",
     "gray-matter": "^4.0.3",
-    "image-size": "^1.0.2",
-    "postcss": "^8.4.23",
-    "prettier": "^2.8.8",
-    "prettier-plugin-tailwindcss": "^0.2.7",
-    "rehype-katex": "^6.0.3",
-    "rehype-pretty-code": "^0.9.5",
-    "rehype-rewrite": "^3.0.6",
-    "remark-directive": "^2.0.1",
-    "remark-gfm": "^3.0.1",
-    "remark-math": "^5.1.1",
-    "remark-unwrap-images": "^3.0.1",
-    "shiki": "^0.14.1",
-    "tailwindcss": "^3.3.2",
+    "image-size": "^1.1.1",
+    "postcss": "^8.4.38",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.14",
+    "rehype-katex": "^7.0.0",
+    "rehype-pretty-code": "^0.13.2",
+    "rehype-rewrite": "^4.0.2",
+    "remark-directive": "^3.0.0",
+    "remark-gfm": "^4.0.0",
+    "remark-math": "^6.0.0",
+    "remark-unwrap-images": "^4.0.0",
+    "shiki": "^1.6.0",
+    "tailwindcss": "^3.4.3",
     "typescript": "^5.0.4",
-    "unist-util-visit": "^4.1.2"
+    "unist-util-visit": "^5.0.0"
   }
 }
