@@ -1,5 +1,5 @@
 import Image, { type ImageProps } from 'next/image';
-import { NextTweet } from 'next-tweet';
+import { Tweet } from 'react-tweet';
 
 import { Callout } from 'components/callout';
 import { Annotation } from 'components/annotation';
@@ -29,9 +29,9 @@ export const mdxComponents = {
   ),
   Annotation,
   Callout,
-  Tweet: (props: Parameters<typeof NextTweet>[0]) => (
+  Tweet: (props: Parameters<typeof Tweet>[0]) => (
     <div className="not-prose">
-      <NextTweet {...props} />
+      <Tweet {...props} />
     </div>
   ),
 };
