@@ -1,7 +1,9 @@
 /** @type {import('postcss').Postcss} */
-module.exports = {
+const config = {
   plugins: {
     tailwindcss: {},
     autoprefixer: {},
   },
 };
+
+export default config;
